+++ packages/node/vitest.config.ts
@@ -1,0 +1,14 @@
+/// <reference types="vitest" />
+
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    testTimeout: 90000,
+  },
+  resolve: {
+    alias: {
+      "@serverless-stack/node": "./src",
+    },
+  },
+});