+++ packages/resources/src/Parameter.ts
@@ -42,7 +42,11 @@ export class Parameter extends Construct {
 
     // Create SSM parameter
     new ssm.StringParameter(this, "Parameter", {
-      parameterName: FunctionConfig.buildSsmNameForParameter(app.name, app.stage, id),
+      parameterName: FunctionConfig.buildSsmNameForParameter(
+        app.name,
+        app.stage,
+        id
+      ),
       stringValue: value,
     });
   }
@@ -70,5 +74,17 @@ export class Parameter extends Construct {
         name: this.name,
       },
     };
+  }
+
+  public static create<T extends Record<string, any>>(
+    scope: Construct,
+    parameters: T
+  ) {
+    const result: Record<string, Parameter> = {};
+    for (const [name, value] of Object.entries(parameters)) {
+      result[name] = new Parameter(scope, name, { value });
+    }
+
+    return result as { [key in keyof T]: Parameter };
   }
 }